Remove `SQLiteHistory`
