--- Storage/SQL/SQLiteHistory.swift
@@ -1,32 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-import Foundation
-import Shared
-import XCGLogger
-
-private let log = Logger.syncLogger
-
-class NoSuchRecordError: MaybeErrorType {
-    let guid: GUID
-    init(guid: GUID) {
-        self.guid = guid
-    }
-    var description: String {
-        return "No such record: \(guid)."
-    }
-}
-
-extension SDRow {
-    func getTimestamp(_ column: String) -> Timestamp? {
-        return (self[column] as? NSNumber)?.uint64Value
-    }
-
-    func getBoolean(_ column: String) -> Bool {
-        if let val = self[column] as? Int {
-            return val != 0
-        }
-        return false
-    }
-}
